+++ src/lazy-collection.ts
@@ -1,10 +1,8 @@
 import { Maybe, None, Some } from '@nkp/maybe';
 import { smartSort, toIterable } from './utils';
 import { $ANY, $TODO } from './utility-types';
-import { Betweenable, Iterateable, Orderable, SortDirection } from './types';
+import { Betweenable, Iterateable, Orderable, SortDirection, Unary } from './types';
 import { ICollection } from './collection.interface';
-
-export interface ForkLazyCollection<T, R> { (collection: LazyCollection<T> ): R; }
 
 
 /**
@@ -114,30 +112,14 @@ export class LazyCollection<T> implements ICollection<T> {
    *
    * @param forks
    */
-  fork<R1>(...forks: readonly [ForkLazyCollection<T, Iterateable<R1>>]): LazyCollection<R1>
-  fork<R1, R2>(...forks: readonly [ForkLazyCollection<T, Iterateable<R1>>, ForkLazyCollection<T, Iterateable<R2>>]): LazyCollection<R1 | R2>
-  fork<R1, R2, R3>(...forks: readonly [ForkLazyCollection<T, Iterateable<R1>>, ForkLazyCollection<T, Iterateable<R2>>, ForkLazyCollection<T, Iterateable<R3>>]): LazyCollection<R1 | R2 | R3>
-  fork<R1, R2, R3, R4>(...forks: readonly [ForkLazyCollection<T, Iterateable<R1>>, ForkLazyCollection<T, Iterateable<R2>>, ForkLazyCollection<T, Iterateable<R3>>, ForkLazyCollection<T, Iterateable<R4>>]): LazyCollection<R1 | R2 | R3 | R4>
-  fork<R1, R2, R3, R4, R5>(...forks: readonly [ForkLazyCollection<T, Iterateable<R1>>, ForkLazyCollection<T, Iterateable<R2>>, ForkLazyCollection<T, Iterateable<R3>>, ForkLazyCollection<T, Iterateable<R4>>, ForkLazyCollection<T, Iterateable<R5>>]): LazyCollection<R1 | R2 | R3 | R4 | R5>
-  fork<R1, R2, R3, R4, R5, R6>(...forks: readonly [ForkLazyCollection<T, Iterateable<R1>>, ForkLazyCollection<T, Iterateable<R2>>, ForkLazyCollection<T, Iterateable<R3>>, ForkLazyCollection<T, Iterateable<R4>>, ForkLazyCollection<T, Iterateable<R5>>, ForkLazyCollection<T, Iterateable<R6>>]): LazyCollection<R1 | R2 | R3 | R4 | R5 | R6>
-  fork<R1, R2, R3, R4, R5, R6, R7>(...forks: readonly [ForkLazyCollection<T, Iterateable<R1>>, ForkLazyCollection<T, Iterateable<R2>>, ForkLazyCollection<T, Iterateable<R3>>, ForkLazyCollection<T, Iterateable<R4>>, ForkLazyCollection<T, Iterateable<R5>>, ForkLazyCollection<T, Iterateable<R6>>, ForkLazyCollection<T, Iterateable<R7>>]): LazyCollection<R1 | R2 | R3 | R4 | R5 | R6 | R7>
-  fork<R1, R2, R3, R4, R5, R6, R7, R8>(...forks: readonly [ForkLazyCollection<T, Iterateable<R1>>, ForkLazyCollection<T, Iterateable<R2>>, ForkLazyCollection<T, Iterateable<R3>>, ForkLazyCollection<T, Iterateable<R4>>, ForkLazyCollection<T, Iterateable<R5>>, ForkLazyCollection<T, Iterateable<R6>>, ForkLazyCollection<T, Iterateable<R7>>, ForkLazyCollection<T, Iterateable<R8>>]): LazyCollection<R1 | R2 | R3 | R4 | R5 | R6 | R7 | R8>
-  fork<R1, R2, R3, R4, R5, R6, R7, R8, R9>(...forks: readonly [ForkLazyCollection<T, Iterateable<R1>>, ForkLazyCollection<T, Iterateable<R2>>, ForkLazyCollection<T, Iterateable<R3>>, ForkLazyCollection<T, Iterateable<R4>>, ForkLazyCollection<T, Iterateable<R5>>, ForkLazyCollection<T, Iterateable<R6>>, ForkLazyCollection<T, Iterateable<R7>>, ForkLazyCollection<T, Iterateable<R8>>, ForkLazyCollection<T, Iterateable<R9>>]): LazyCollection<R1 | R2 | R3 | R4 | R5 | R6 | R7 | R8 | R9>
-  fork<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10>(...forks: readonly [ForkLazyCollection<T, Iterateable<R1>>, ForkLazyCollection<T, Iterateable<R2>>, ForkLazyCollection<T, Iterateable<R3>>, ForkLazyCollection<T, Iterateable<R4>>, ForkLazyCollection<T, Iterateable<R5>>, ForkLazyCollection<T, Iterateable<R6>>, ForkLazyCollection<T, Iterateable<R7>>, ForkLazyCollection<T, Iterateable<R8>>, ForkLazyCollection<T, Iterateable<R9>>, ForkLazyCollection<T, Iterateable<R10>>]): LazyCollection<R1 | R2 | R3 | R4 | R5 | R6 | R7 | R8 | R9 | R10>
-  fork<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11>(...forks: readonly [ForkLazyCollection<T, Iterateable<R1>>, ForkLazyCollection<T, Iterateable<R2>>, ForkLazyCollection<T, Iterateable<R3>>, ForkLazyCollection<T, Iterateable<R4>>, ForkLazyCollection<T, Iterateable<R5>>, ForkLazyCollection<T, Iterateable<R6>>, ForkLazyCollection<T, Iterateable<R7>>, ForkLazyCollection<T, Iterateable<R8>>, ForkLazyCollection<T, Iterateable<R9>>, ForkLazyCollection<T, Iterateable<R10>>, ForkLazyCollection<T, Iterateable<R11>>]): LazyCollection<R1 | R2 | R3 | R4 | R5 | R6 | R7 | R8 | R9 | R10 | R11>
-  fork<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12>(...forks: readonly [ForkLazyCollection<T, Iterateable<R1>>, ForkLazyCollection<T, Iterateable<R2>>, ForkLazyCollection<T, Iterateable<R3>>, ForkLazyCollection<T, Iterateable<R4>>, ForkLazyCollection<T, Iterateable<R5>>, ForkLazyCollection<T, Iterateable<R6>>, ForkLazyCollection<T, Iterateable<R7>>, ForkLazyCollection<T, Iterateable<R8>>, ForkLazyCollection<T, Iterateable<R9>>, ForkLazyCollection<T, Iterateable<R10>>, ForkLazyCollection<T, Iterateable<R11>>, ForkLazyCollection<T, Iterateable<R12>>]): LazyCollection<R1 | R2 | R3 | R4 | R5 | R6 | R7 | R8 | R9 | R10 | R11 | R12>
-  fork<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13>(...forks: readonly [ForkLazyCollection<T, Iterateable<R1>>, ForkLazyCollection<T, Iterateable<R2>>, ForkLazyCollection<T, Iterateable<R3>>, ForkLazyCollection<T, Iterateable<R4>>, ForkLazyCollection<T, Iterateable<R5>>, ForkLazyCollection<T, Iterateable<R6>>, ForkLazyCollection<T, Iterateable<R7>>, ForkLazyCollection<T, Iterateable<R8>>, ForkLazyCollection<T, Iterateable<R9>>, ForkLazyCollection<T, Iterateable<R10>>, ForkLazyCollection<T, Iterateable<R11>>, ForkLazyCollection<T, Iterateable<R12>>, ForkLazyCollection<T, Iterateable<R13>>]): LazyCollection<R1 | R2 | R3 | R4 | R5 | R6 | R7 | R8 | R9 | R10 | R11 | R12 | R13>
-  fork<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13, R14>(...forks: readonly [ForkLazyCollection<T, Iterateable<R1>>, ForkLazyCollection<T, Iterateable<R2>>, ForkLazyCollection<T, Iterateable<R3>>, ForkLazyCollection<T, Iterateable<R4>>, ForkLazyCollection<T, Iterateable<R5>>, ForkLazyCollection<T, Iterateable<R6>>, ForkLazyCollection<T, Iterateable<R7>>, ForkLazyCollection<T, Iterateable<R8>>, ForkLazyCollection<T, Iterateable<R9>>, ForkLazyCollection<T, Iterateable<R10>>, ForkLazyCollection<T, Iterateable<R11>>, ForkLazyCollection<T, Iterateable<R12>>, ForkLazyCollection<T, Iterateable<R13>>, ForkLazyCollection<T, Iterateable<R14>>]): LazyCollection<R1 | R2 | R3 | R4 | R5 | R6 | R7 | R8 | R9 | R10 | R11 | R12 | R13 | R14>
-  fork<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13, R14, R15>(...forks: readonly [ForkLazyCollection<T, Iterateable<R1>>, ForkLazyCollection<T, Iterateable<R2>>, ForkLazyCollection<T, Iterateable<R3>>, ForkLazyCollection<T, Iterateable<R4>>, ForkLazyCollection<T, Iterateable<R5>>, ForkLazyCollection<T, Iterateable<R6>>, ForkLazyCollection<T, Iterateable<R7>>, ForkLazyCollection<T, Iterateable<R8>>, ForkLazyCollection<T, Iterateable<R9>>, ForkLazyCollection<T, Iterateable<R10>>, ForkLazyCollection<T, Iterateable<R11>>, ForkLazyCollection<T, Iterateable<R12>>, ForkLazyCollection<T, Iterateable<R13>>, ForkLazyCollection<T, Iterateable<R14>>, ForkLazyCollection<T, Iterateable<R15>>]): LazyCollection<R1 | R2 | R3 | R4 | R5 | R6 | R7 | R8 | R9 | R10 | R11 | R12 | R13 | R14 | R15>
-  fork<R>(...forks: readonly (ForkLazyCollection<T, Iterateable<R>>)[]): LazyCollection<R>
-  fork<R>(...forks: readonly (ForkLazyCollection<T, Iterateable<R>>)[]): LazyCollection<R> {
+  forkFlat<R>(...forks: readonly (Unary<this, Iterateable<R>>)[]): LazyCollection<R> {
     const combined = this
       // split forks
-      .split(...forks)
-      // collect results
-      .map((splits) => splits.map(stream => new LazyCollection(stream)))
-      // join results
+      .forkMap(...forks)
       .flat()
+      // normalize
+      .map((splits) => new LazyCollection(splits))
+      // join results
       .flat();
     return combined;
   }
@@ -147,28 +129,52 @@ export class LazyCollection<T> implements ICollection<T> {
    *
    * Similar to Promise.all
    */
-  split<R1>(...splits: readonly [ForkLazyCollection<T, R1>]): LazyCollection<[R1]>
-  split<R1, R2>(...splits: readonly [ForkLazyCollection<T, R1>, ForkLazyCollection<T, R2>]): LazyCollection<[R1, R2]>
-  split<R1, R2, R3>(...splits: readonly [ForkLazyCollection<T, R1>, ForkLazyCollection<T, R2>, ForkLazyCollection<T, R3>]): LazyCollection<[R1, R2, R3]>
-  split<R1, R2, R3, R4>(...splits: readonly [ForkLazyCollection<T, R1>, ForkLazyCollection<T, R2>, ForkLazyCollection<T, R3>, ForkLazyCollection<T, R4>]): LazyCollection<[R1, R2, R3, R4]>
-  split<R1, R2, R3, R4, R5>(...splits: readonly [ForkLazyCollection<T, R1>, ForkLazyCollection<T, R2>, ForkLazyCollection<T, R3>, ForkLazyCollection<T, R4>, ForkLazyCollection<T, R5>]): LazyCollection<[R1, R2, R3, R4, R5]>
-  split<R1, R2, R3, R4, R5, R6>(...splits: readonly [ForkLazyCollection<T, R1>, ForkLazyCollection<T, R2>, ForkLazyCollection<T, R3>, ForkLazyCollection<T, R4>, ForkLazyCollection<T, R5>, ForkLazyCollection<T, R6>]): LazyCollection<[R1, R2, R3, R4, R5, R6]>
-  split<R1, R2, R3, R4, R5, R6, R7>(...splits: readonly [ForkLazyCollection<T, R1>, ForkLazyCollection<T, R2>, ForkLazyCollection<T, R3>, ForkLazyCollection<T, R4>, ForkLazyCollection<T, R5>, ForkLazyCollection<T, R6>, ForkLazyCollection<T, R7>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7]>
-  split<R1, R2, R3, R4, R5, R6, R7, R8>(...splits: readonly [ForkLazyCollection<T, R1>, ForkLazyCollection<T, R2>, ForkLazyCollection<T, R3>, ForkLazyCollection<T, R4>, ForkLazyCollection<T, R5>, ForkLazyCollection<T, R6>, ForkLazyCollection<T, R7>, ForkLazyCollection<T, R8>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8]>
-  split<R1, R2, R3, R4, R5, R6, R7, R8, R9>(...splits: readonly [ForkLazyCollection<T, R1>, ForkLazyCollection<T, R2>, ForkLazyCollection<T, R3>, ForkLazyCollection<T, R4>, ForkLazyCollection<T, R5>, ForkLazyCollection<T, R6>, ForkLazyCollection<T, R7>, ForkLazyCollection<T, R8>, ForkLazyCollection<T, R9>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9]>
-  split<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10>(...splits: readonly [ForkLazyCollection<T, R1>, ForkLazyCollection<T, R2>, ForkLazyCollection<T, R3>, ForkLazyCollection<T, R4>, ForkLazyCollection<T, R5>, ForkLazyCollection<T, R6>, ForkLazyCollection<T, R7>, ForkLazyCollection<T, R8>, ForkLazyCollection<T, R9>, ForkLazyCollection<T, R10>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9, R10]>
-  split<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11>(...splits: readonly [ForkLazyCollection<T, R1>, ForkLazyCollection<T, R2>, ForkLazyCollection<T, R3>, ForkLazyCollection<T, R4>, ForkLazyCollection<T, R5>, ForkLazyCollection<T, R6>, ForkLazyCollection<T, R7>, ForkLazyCollection<T, R8>, ForkLazyCollection<T, R9>, ForkLazyCollection<T, R10>, ForkLazyCollection<T, R11>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11]>
-  split<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12>(...splits: readonly [ForkLazyCollection<T, R1>, ForkLazyCollection<T, R2>, ForkLazyCollection<T, R3>, ForkLazyCollection<T, R4>, ForkLazyCollection<T, R5>, ForkLazyCollection<T, R6>, ForkLazyCollection<T, R7>, ForkLazyCollection<T, R8>, ForkLazyCollection<T, R9>, ForkLazyCollection<T, R10>, ForkLazyCollection<T, R11>, ForkLazyCollection<T, R12>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12]>
-  split<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13>(...splits: readonly [ForkLazyCollection<T, R1>, ForkLazyCollection<T, R2>, ForkLazyCollection<T, R3>, ForkLazyCollection<T, R4>, ForkLazyCollection<T, R5>, ForkLazyCollection<T, R6>, ForkLazyCollection<T, R7>, ForkLazyCollection<T, R8>, ForkLazyCollection<T, R9>, ForkLazyCollection<T, R10>, ForkLazyCollection<T, R11>, ForkLazyCollection<T, R12>, ForkLazyCollection<T, R13>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13]>
-  split<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13, R14>(...splits: readonly [ForkLazyCollection<T, R1>, ForkLazyCollection<T, R2>, ForkLazyCollection<T, R3>, ForkLazyCollection<T, R4>, ForkLazyCollection<T, R5>, ForkLazyCollection<T, R6>, ForkLazyCollection<T, R7>, ForkLazyCollection<T, R8>, ForkLazyCollection<T, R9>, ForkLazyCollection<T, R10>, ForkLazyCollection<T, R11>, ForkLazyCollection<T, R12>, ForkLazyCollection<T, R13>, ForkLazyCollection<T, R14>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13, R14]>
-  split<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13, R14, R15>(...splits: readonly [ForkLazyCollection<T, R1>, ForkLazyCollection<T, R2>, ForkLazyCollection<T, R3>, ForkLazyCollection<T, R4>, ForkLazyCollection<T, R5>, ForkLazyCollection<T, R6>, ForkLazyCollection<T, R7>, ForkLazyCollection<T, R8>, ForkLazyCollection<T, R9>, ForkLazyCollection<T, R10>, ForkLazyCollection<T, R11>, ForkLazyCollection<T, R12>, ForkLazyCollection<T, R13>, ForkLazyCollection<T, R14>, ForkLazyCollection<T, R15>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13, R14, R15]>
-  split<R>(...splits: readonly (ForkLazyCollection<T, R>)[]): LazyCollection<R[]>
-  split<R>(...splits: readonly (ForkLazyCollection<T, R>)[]): LazyCollection<R[]> {
+  forkMap<M extends Record<PropertyKey, Unary<this, unknown>>>(forks: M): LazyCollection<{ [K in keyof M]: ReturnType<M[K]> }>;
+  forkMap<R1>(...splits: readonly [Unary<this, R1>]): LazyCollection<[R1]>
+  forkMap<R1, R2>(...splits: readonly [Unary<this, R1>, Unary<this, R2>]): LazyCollection<[R1, R2]>
+  forkMap<R1, R2, R3>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>]): LazyCollection<[R1, R2, R3]>
+  forkMap<R1, R2, R3, R4>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>]): LazyCollection<[R1, R2, R3, R4]>
+  forkMap<R1, R2, R3, R4, R5>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>]): LazyCollection<[R1, R2, R3, R4, R5]>
+  forkMap<R1, R2, R3, R4, R5, R6>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>]): LazyCollection<[R1, R2, R3, R4, R5, R6]>
+  forkMap<R1, R2, R3, R4, R5, R6, R7>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7]>
+  forkMap<R1, R2, R3, R4, R5, R6, R7, R8>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>, Unary<this, R8>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8]>
+  forkMap<R1, R2, R3, R4, R5, R6, R7, R8, R9>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>, Unary<this, R8>, Unary<this, R9>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9]>
+  forkMap<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>, Unary<this, R8>, Unary<this, R9>, Unary<this, R10>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9, R10]>
+  forkMap<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>, Unary<this, R8>, Unary<this, R9>, Unary<this, R10>, Unary<this, R11>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11]>
+  forkMap<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>, Unary<this, R8>, Unary<this, R9>, Unary<this, R10>, Unary<this, R11>, Unary<this, R12>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12]>
+  forkMap<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>, Unary<this, R8>, Unary<this, R9>, Unary<this, R10>, Unary<this, R11>, Unary<this, R12>, Unary<this, R13>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13]>
+  forkMap<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13, R14>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>, Unary<this, R8>, Unary<this, R9>, Unary<this, R10>, Unary<this, R11>, Unary<this, R12>, Unary<this, R13>, Unary<this, R14>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13, R14]>
+  forkMap<R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13, R14, R15>(...splits: readonly [Unary<this, R1>, Unary<this, R2>, Unary<this, R3>, Unary<this, R4>, Unary<this, R5>, Unary<this, R6>, Unary<this, R7>, Unary<this, R8>, Unary<this, R9>, Unary<this, R10>, Unary<this, R11>, Unary<this, R12>, Unary<this, R13>, Unary<this, R14>, Unary<this, R15>]): LazyCollection<[R1, R2, R3, R4, R5, R6, R7, R8, R9, R10, R11, R12, R13, R14, R15]>
+  forkMap<R>(...splits: readonly (Unary<this, R>)[]): LazyCollection<R[]>
+  forkMap(mapOrFunction?: (Unary<this, unknown> | Record<PropertyKey, Unary<this, unknown>>), ...rest: Unary<this, unknown>[]): $ANY {
     const self = this;
-    function * iterateable(): Iterable<R[]> {
-      yield splits.map(split => split(self));
+    if (mapOrFunction && !Array.isArray(mapOrFunction)) {
+      const mapping = mapOrFunction as Record<PropertyKey, Unary<this, unknown>>;
+      // given a map
+      const iterateable = function * iterateable(): Iterable<Record<string, unknown>> {
+        const mappedResults: Record<PropertyKey, $ANY> = {};
+        Object.keys(mapping).map(key => {
+          mappedResults[key] = mapping[key]!(self);
+        });
+        yield mappedResults;
+      };
+      return new LazyCollection(iterateable);
     }
-    return new LazyCollection(iterateable);
+
+    else if (mapOrFunction) {
+      const zeroFn = mapOrFunction as Unary<this, unknown>;
+      // given an array of functions
+      // given an array
+      const fns = [zeroFn, ...rest,] as Unary<this, unknown>[];
+      const iterateable = function * iterateable(): Iterable<unknown[]> {
+        const mapResults = fns.map((split) => split(self));
+        yield mapResults;
+      };
+      return new LazyCollection(iterateable);
+    }
+
+    // given nothing
+    return new LazyCollection([]);
   }